now all controllers exists

+++ src/main/java/com/sedbproj/Railways/Repositories/CarriageRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface CarriageRepository extends CrudRepository<CarriageEntity, Long> {
+public interface CarriageRepository extends CrudRepository<CarriageEntity, Integer> {
     List<CarriageEntity> getCarriageEntitiesByTrainId(Integer id);
 }
